python/requirements.txt
pyserial==3.5
python-uinput==1.0.1
